fix enabled services in controller node
docker and kubelet are services that should be running in the controller
node. However they are disabled by default because they are not meant to
be running in the "worker" nodes.

salt-minion is the opposite case. salt-minion should be running in all
nodes except in the controller node, thus we should disable it.