Remove hardcoded secrets
We will be generating secrets with init containers. These secrets will
be created in a volume mounted from the host, so they survive reboots.

While being sufficient for our GA purposes we will need to rethink how
we do this in a HA environment.

Some secrets are generated with the init containers:

* mysql root password
* mysql velum user password
* mysql salt user password
* saltapi user password

Once we have generated all the passwords, we need to write this
configuration on files that will be mounted on the different containers,
so the different services can read the files where the passwords are written.
By default, passwords will be created in files with permissions 400.

Password generation uses `/dev/random`, performing a `base64` encoding to
that random content, and pick up a line of the `base64` output.

Images will take this environment variables and they will use their
entrypoint to perform the required actions. Example:

* mariadb container will set the root password and do some initializations
* salt-master container will `chpasswd` the `saltapi` user to the generated
  saltapi password.