Make substitution in a safer way for --pod-infra-container-image argument
This wasn't working on our production image because we are using Kubernetes 1.5
that in our config comes with the following setting in /etc/kubernetes/kubelet:

KUBELET_ARGS="--config=/etc/kubernetes/manifests"

On 1.6, --config has been completely removed and it will use --pod-manifest-path,
but not on our current installed configuration.

By adding this change, we ensure that we only make the replacement once (if the
pod-manifest-path is already there we won't do anything), and we don't rely on the
current contents for making the substitution.

Fixes: bsc#1039863