Do not mount `/usr/share/salt/kubernetes/config/master.d` from the host
We will mounting other volumes on top of this on the containers, and
they will fail because on the host, `/usr/share/salt/kubernetes/config/master.d`
is a `RO` volume.

We fix this by mounting all specific files in the containers instead of
the top level directory of the hierarchy. This imposes us the restriction
to modify the container manifests every time a new config file appears, but
that should not happen very often.

Otherwise, we cannot add our own configuration files on top of the `RO`
mounted volume, because they will fail. In this case, the mounted folder
on the containers will be `/etc/salt/master.d`, but in this case this folder
won't be mounted from `/usr/share/salt/kubernetes/config/master.d`, it will
live only in the container, and we will mount the specific files under it,
what will avoid the `RO` volume problems from the host.