bsc#1043592 Add pre-generation of minion keys
Generates 2 salt keys (ca and admin) and places them in the correct
directories.

This allows us to remove "auto_accept" from the master config file
and select the rest of the members of the cluster.

The admin key is writen out to `/etc/salt/pki/minion/minion.(pub|pem)`
The ca key is written out the same path in the container.

bsc#1043592