Add `cert-init.yaml` manifest that will initialize certificates
We will initialize certificates using salt (velum and salt-api). This
new manifest will include a container that will wait for basic
initialization to happen on the `public.yaml` manifest, and then it
will signal an event to the salt-master, so it can run an orchestration
and generate the certificates for the very first services (velum and
salt-api for now).

This will only happen once, as this event will only be triggered if
the certificates aren't there yet.

Velum and salt-api services will refer to the certificates, so they
will retry to start until the certificates are present.

Fixes: bsc#1043570
Fixes: bsc#1043589