bsc#1045350 Accept salt keys that have been pre-generated
Currently the admin nodes salt minion starts before the container
that generates and accepts keys is ran.

This means that the salt minion is started with a key that is not
accepted, and goes to a pending state.

This checks if the key is pre-generated, and if we have accepted
a key from this minion before. If the key has been generated,
but not accepted, we accept the key and continue.