Rename `velum-dashboard-autoyast` to `velum-autoyast`
We have a lot of processes in the development, e2e-tests and debugging
environments that use `velum-dashboard`. Renaming the autoyast serving
to `velum-autoyast` will make them still only match one container, the
one they expect (actually both of them are practically the same thing,
but to keep things as they were).