Add kubic images
If this package is built in tumbleweed,the images should be named
tumbleweed and not sle12.

When this builds on sle15, the images should be named sle15.

We cannot use the same name for the different suse versions as if this
images were in a registry, they should be named different.