Make entrypoint of mariadb-user-secrets container more robust
I've run into a timing issue that caused the root password of mariadb
**not** being injected into the running container "mariadb-user-secrets"
in time. That caused the container to enter an infinite loop consisting
of trying to connect to mariadb as root without a specifying password,
getting an error message, sleeping 1 second and trying again.

This is an init container, as long as it's running kubelet won't start
over containers, like openldap, velum-*, salt-*,...

With this change the mariadb entrypoint waits untile the file containing
the root password exists and is not empty.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>