Run Velum's `bin/init` as an init container
This will ensure that when the dashboard, api and event processor start,
the database has been created and/or migrated at all times. This avoids
the weird situation in which the api and event-processor can start while
the dashboard was still migrating the database, causing side-effects if
Rails already cached some model attributes on other processes.

Fixes: bsc#1091843