Configure `innodb_log_file_size` to a 128M limit
When a salt output is big enough, mysql will refuse to insert the offending
row for being too big, with an error:

```
[ERROR   ] Could not store events - returner 'mysql.event_return' raised exception:
(1118, 'The size of BLOB/TEXT data inserted in one transaction is greater than 10% of redo log size.
Increase the redo log size using innodb_log_file_size.')
```

Whatever we set as `innodb_log_file_size` will be an arbitrary number that will
eventually be flooded if the cluster is big enough, or if salt is noisy enough.
Given a cluster size, this can suddenly fail if we add more states (thus, increasing
salt's output). Obviously, given the same salt states, we can also reach this limit
by increasing the cluster size.

There is not a definitive fix for this issue, all we can do for now (without a proper
refactor of the way we integrate salt and velum) is to ensure that with todays salt
states we can reach a certain number of nodes. As said, this can no longer be true if
we add more salt states and we reach again the limit for the same cluster size.

Fixes: bsc#1095335