Do not mount the public directory or the compiled assets get hidden in the iso
Mounting the public directory from the host, was "hidding" the assets
files present in the public directory and so our images were broken.

This was only happening in the iso and not in our dev environments
because in production we have the assets precompiled and we don't
compile them on the fly, since we don't have a js engine installed.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>