use the sles12 images from registry.suse.de
This is a first step towards using the registry.
We will use an internal one in registry.suse.de
Later, we will use the registry.suse.com for caasp and
registry.opensuse.org for kubic, however, we have to figure out the
tagging strategy if we want to use kubic images or images based on
sle15.

Thus, for now, we start using the sles12 images which have a fixed tag.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>