Map whole /var/lib/ca-certificates/ directory
Because updating the ca-bundle .pem file creates a new file instead of
truncating and overwriting (which is the correct way to do things), the
bind mount of /var/lib/ca-certificates/ca-bundle.pem breaks after
update-ca-certificates is run.  Mount the whole directory into the
containers instead, since there's nothing else in that directory which
should differ anyway.  This also gets the OpenSSL bundle directory (which
is a symlink targeted in this dir) mapped in for free.