Set password for DB users separate from creation
If a user already exists but teh secret file has been modified, the database
user's password will not be changed for an existing user.  Further, the create
user command will generate an error if run against a user which already exists.
Separate out the user create and password set commands so the create can add
the "if not exsts" option and so the password can be resync'd with the file
contents.